TVIST1-359: Moved reminder new, complete and edit to CaseReminderController
