+++ templates/case/index.html.twig
@@ -89,7 +89,7 @@
                             </div>
                         </td>
                     </tr>
-                    {{ render(controller('App\\Controller\\ReminderController:new', {id: case.id})) }}
+                    {{ render(controller('App\\Controller\\CaseReminderController:new', {id: case.id})) }}
                 {% endfor %}
                 </tbody>
             </table>
