+++ templates/reminder/index.html.twig
@@ -35,13 +35,18 @@
                         <td>{{ reminder.content|e }}</td>
                         <td>
                             <div class="btn-group" role="group" aria-label="{% trans %}List functions{% endtrans %}">
-                                <a href="{{ url('reminder_complete', {id: reminder.id}) }}" role="button" onclick="return confirm('{% trans %}Are you sure you want to complete this reminder?{% endtrans %}');" class="btn btn-success" data-toggle="tooltip" data-placement="auto" title="{% trans %}Mark as completed{% endtrans %}"><i class="fas fa-check-circle"></i><span class="sr-only">{% trans %}Mark as completed{% endtrans %} </span></a>
+                                <a href="{{ url('reminder_complete', {id: reminder.caseEntity.id, reminder_id: reminder.id}) }}" role="button" onclick="return confirm('{% trans %}Are you sure you want to complete this reminder?{% endtrans %}');" class="btn btn-success" data-toggle="tooltip" data-placement="auto" title="{% trans %}Mark as completed{% endtrans %}"><i class="fas fa-check-circle"></i><span class="sr-only">{% trans %}Mark as completed{% endtrans %} </span></a>
                                 <a href="" role="button" class="btn btn-warning" data-toggle="modal" data-target="#reminder_edit{{ reminder.id }}" data-placement="auto" title="{% trans %}Edit reminder{% endtrans %}"><i class="fas fa-pen"></i><span class="sr-only">{% trans %}Edit reminder{% endtrans %}</span></a>
                                 <a href="{{ url('case_summary', {id: reminder.caseEntity.id}) }}" class="btn btn-primary" role="button" data-toggle="tooltip" data-placement="auto" title="{% trans %}Show case{% endtrans %} {{ reminder.caseEntity.casenumber|e }}"><i class="fas fa-file-alt"></i><span class="sr-only">"{% trans %}Show case{% endtrans %} {{ reminder.caseEntity.casenumber|e }}"</span></a>
                             </div>
                         </td>
                     </tr>
-                    {{ render(controller('App\\Controller\\ReminderController:edit', {id: reminder.id})) }}
+                    {{ render(controller(
+                        'App\\Controller\\CaseReminderController:edit', {
+                            id: reminder.caseEntity.id,
+                            reminder_id: reminder.id
+                            }
+                        )) }}
                 {% endfor %}
                 </tbody>
             </table>
